+++ volunteer/templates/admin/login.html
@@ -4,6 +4,6 @@
 {% block content %}
 <h1>Volunteer Database Admin Login</h1>
     <section>    	       
-        <p>This is a sub page for the site admin that doesn't exist yet</p>
+        <p>Nothing here.  Use the normal <a href="{% url 'login' %}">Login</a> page.</p>
     </section>
 {% endblock %}
